pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coral-thesis"
version = "0.1.0"
description = "Clean V2 coral health assessment pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy<2.0",
  "opencv-python",
  "pandas",
  "PyYAML",
  "scikit-learn",
  "ultralytics",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
coral-thesis = "coral_thesis.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

